move xmlsql, and change directory xmlTag from Rest to Xml
